+++ include/build_info.h
@@ -1,4 +1,4 @@
 #pragma once
 
 constexpr const char* BUILD_ID =
-    "7IN-20260729-PRODUCT53R5-AIRPORT-EYE";
+    "7IN-20260731-PRODUCT53R6-AIRPORT-TOUCH";